+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/FilteredTable.java
@@ -455,10 +455,21 @@ public final class FilteredTable<R, C> extends JTable {
 
 	/**
 	 * Copies the table data as a TAB delimited string, with header, to the clipboard.
-	 * Note that if the selection is empty all rows are copied, otherwise only selected rows.
 	 */
-	public void copyRowsAsTabDelimitedString() {
-		Utilities.setClipboard(tableModel.rowsAsDelimitedString('\t'));
+	public void copyToClipboard() {
+		Utilities.setClipboard(tableModel.export()
+						.delimiter('\t')
+						.get());
+	}
+
+	/**
+	 * Copies the table data as a TAB delimited string, with header, to the clipboard.
+	 */
+	public void copySelectedToClipboard() {
+		Utilities.setClipboard(tableModel.export()
+						.delimiter('\t')
+						.selected(true)
+						.get());
 	}
 
 	/**